tests: update to not run vgconvert -M with lvmetad

+++ test/shell/pvcreate-usage.sh
@@ -131,18 +131,30 @@ check pv_field "$dev1" pv_mda_count 2
 
 #COMM 'pv with LVM1 compatible data alignment can be convereted'
 #compatible == LVM1_PE_ALIGN == 64k
+if test ! -e LOCAL_LVMETAD; then
 pvcreate --dataalignment 256k "$dev1"
 vgcreate -s 1m $vg "$dev1"
 vgconvert -M1 $vg
 vgconvert -M2 $vg
 check pv_field "$dev1" pe_start 256.00k
 vgremove $vg
+fi
 
 #COMM 'pv with LVM1 incompatible data alignment cannot be convereted'
+if test ! -e LOCAL_LVMETAD; then
 pvcreate --dataalignment 10k "$dev1"
 vgcreate -s 1m $vg "$dev1"
 not vgconvert -M1 $vg
 vgremove $vg
+fi
+
+#COMM 'vgconvert -M is disallowed with lvmetad'
+if test -e LOCAL_LVMETAD; then
+pvcreate "$dev1"
+vgcreate $vg "$dev1"
+not vgconvert -M1 $vg
+vgremove $vg
+fi
 
 #COMM 'vgcfgrestore allows pe_start=0'
 #basically it produces nonsense, but it tests vgcfgrestore,
